+++ examples/iam-role-for-service-accounts-eks/main.tf
@@ -25,6 +25,7 @@ locals {
 module "disabled" {
   source = "../../modules/iam-role-for-service-accounts-eks"
 
+  role_name   = "disabled"
   create_role = false
 }
 
@@ -352,6 +353,22 @@ module "vpc_cni_ipv6_irsa_role" {
   tags = local.tags
 }
 
+module "cloudwatch_observability_irsa_role" {
+  source = "../../modules/iam-role-for-service-accounts-eks"
+
+  role_name                              = "cloudwatch-observability"
+  attach_cloudwatch_observability_policy = true
+
+  oidc_providers = {
+    ex = {
+      provider_arn               = module.eks.oidc_provider_arn
+      namespace_service_accounts = ["amazon-cloudwatch:cloudwatch-agent"]
+    }
+  }
+
+  tags = local.tags
+}
+
 ################################################################################
 # Custom IRSA Roles
 ################################################################################
@@ -405,7 +422,7 @@ module "iam_eks_role" {
 
 module "vpc" {
   source  = "terraform-aws-modules/vpc/aws"
-  version = "~> 4.0"
+  version = "~> 5.0"
 
   name = local.name
   cidr = local.vpc_cidr
@@ -430,10 +447,10 @@ module "vpc" {
 
 module "eks" {
   source  = "terraform-aws-modules/eks/aws"
-  version = "~> 19.14"
+  version = "~> 19.21"
 
   cluster_name    = local.name
-  cluster_version = "1.26"
+  cluster_version = "1.28"
 
   vpc_id     = module.vpc.vpc_id
   subnet_ids = module.vpc.private_subnets