+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All notable changes to this project will be documented in this file.
 
+## [5.33.0](https://github.com/terraform-aws-modules/terraform-aws-iam/compare/v5.32.1...v5.33.0) (2023-12-16)
+
+
+### Features
+
+* Add support for Amazon CloudWatch Observability IRSA role ([#446](https://github.com/terraform-aws-modules/terraform-aws-iam/issues/446)) ([25e2bf9](https://github.com/terraform-aws-modules/terraform-aws-iam/commit/25e2bf9f9f4757a7014b55db981be9d2beeab445))
+
+### [5.32.1](https://github.com/terraform-aws-modules/terraform-aws-iam/compare/v5.32.0...v5.32.1) (2023-12-11)
+
+
+### Bug Fixes
+
+* Remove unused TLS provider in `iam-github-oidc-role` ([#439](https://github.com/terraform-aws-modules/terraform-aws-iam/issues/439)) ([2ce3885](https://github.com/terraform-aws-modules/terraform-aws-iam/commit/2ce3885a6d511a6512a9607abcf839939bbf7b59))
+
+## [5.32.0](https://github.com/terraform-aws-modules/terraform-aws-iam/compare/v5.31.0...v5.32.0) (2023-11-17)
+
+
+### Features
+
+* Add instance profile permissions to Karpenter IRSA policy ([#434](https://github.com/terraform-aws-modules/terraform-aws-iam/issues/434)) ([50348dd](https://github.com/terraform-aws-modules/terraform-aws-iam/commit/50348dd6ac008a38fce1c8d3549915c1dd9ccb65)), closes [#433](https://github.com/terraform-aws-modules/terraform-aws-iam/issues/433)
+
 ## [5.31.0](https://github.com/terraform-aws-modules/terraform-aws-iam/compare/v5.30.2...v5.31.0) (2023-11-13)